small name refactor; few tests;
