+++ workers/lite_worker.py
